Adapt external resources panel to the new dashboard
