+++ sematic/ui/packages/common/src/hooks/externalResourceHooks.ts
@@ -1,5 +1,5 @@
-import { ExternalResource, ExternalResourceState, Run } from "@sematic/common/src/Models";
-import { useHttpClient } from "@sematic/common/src/hooks/httpHooks";
+import { ExternalResource, ExternalResourceState, Run } from "src/Models";
+import { useHttpClient } from "src/hooks/httpHooks";
 import { useEffect, useRef } from "react";
 import useAsyncRetry from "react-use/lib/useAsyncRetry";
 import usePreviousDistinct from "react-use/lib/usePreviousDistinct";
