+++ sematic/ui/packages/common/src/pages/RunDetails/RunTabs.tsx
@@ -9,6 +9,7 @@ import InputPane from "src/pages/RunDetails/artifacts/InputPane";
 import OutputPane from "src/pages/RunDetails/artifacts/OutputPane";
 import LogsPane from "src/pages/RunDetails/logs/LogsPane";
 import RunMetricsPanel from "src/pages/RunDetails/metrics";
+import ExternalResourcePanel from "src/pages/RunDetails/externalResource";
 import theme from "src/theme/new";
 import { useAtom } from "jotai";
 
@@ -35,9 +36,11 @@ const StyledTabPanel = styled(TabPanel)`
     overflow-x: hidden;
     overflow-y: auto;
     scrollbar-gutter: stable;
+    padding-top: ${theme.spacing(5)};
 `;
 
 const StyledTabPanelWithoutMargin = styled(StyledTabPanel)`
+    padding-top: 0;
     margin-left: -${theme.spacing(5)};
     margin-right: -${theme.spacing(5)};
 `;
@@ -85,9 +88,9 @@ const RunTabs = (props: RunTabsProps) => {
         <StyledTabPanel value="metrics">
             <RunMetricsPanel />
         </StyledTabPanel>
-        <TabPanel value="ext_res">
-            <div />
-        </TabPanel>
+        <StyledTabPanel value="ext_res">
+            <ExternalResourcePanel />
+        </StyledTabPanel>
     </TabContext>
 };
